fix(bench-report-ui): production define, dead payload, shared html helpers, node-click selection reset
- vite: define process.env.NODE_ENV=production so the IIFE ships prod React only and no longer references process at file:// mount
- add bundleRuntime test executing each dist bundle in a process-less jsdom realm and guarding against raw </script> in the JS
- CaseDetail: clear trade/action selection on process-node click so snapshot navigation stops fighting the selected trade
- drop dead direction/netR from the chart trade ref type, builder, and fixtures
- consolidate escapeHtml/serializeForScript into report/htmlFormat and import from both renderers
- extract shared report view-data fixtures for reuse across component and runtime tests
