+++ packages/bench-report-ui/test/fixtures.ts
@@ -1,0 +1,296 @@
+import type {
+  EpisodeReportCaseDetailView,
+  EpisodeReportCaseRowView,
+  EpisodeReportChartPayload,
+  EpisodeReportViewData,
+  LeaderboardReportViewData,
+} from '../src/types';
+
+function caseRow(index: number, model: string): EpisodeReportCaseRowView {
+  return {
+    index,
+    anchorId: `case-${index}`,
+    symbol: `SYM${index}`,
+    provenanceSymbol: null,
+    provenanceDate: null,
+    questionId: `Q${index}`,
+    model,
+    mode: 'blind',
+    modeLabel: '盲盘',
+    rep: 1,
+    direction: 'long',
+    directionLabel: '做多',
+    firstDecisionLabel: 'B0 首次决策',
+    planEntry: 100,
+    planStop: 90,
+    planTarget: 120,
+    actualEntry: 100,
+    actualExit: 110,
+    tradeCount: 1,
+    outcome: 'target',
+    outcomeLabel: '止盈',
+    netR: 1,
+    tone: 'positive',
+    mfeR: 1,
+    maeR: 0,
+    costUsd: 0.1,
+    durationLabel: '1.0 s',
+    filterSearch: `sym${index} q${index}`,
+  };
+}
+
+function chartPayload(index: number): EpisodeReportChartPayload {
+  const bars = Array.from({ length: 22 }, (_, i) => ({
+    time: i + 1,
+    open: 100 + i,
+    high: 101 + i,
+    low: 99 + i,
+    close: 100 + i,
+    volume: 100,
+  }));
+  return {
+    id: `trade-chart-${index}`,
+    symbol: `SYM${index}`,
+    finalBarIndex: 0,
+    baseRanges: { h1: bars, day: bars, week: bars },
+    replayH1: [],
+    snapshotPatches: {},
+    markers: { h1: [], day: [], week: [] },
+    levels: [
+      { title: '计划入场', price: 100, color: '#2563eb' },
+      { title: '止损', price: 90, color: '#dc2626' },
+      { title: '止盈', price: 120, color: '#059669' },
+    ],
+    trades: [
+      {
+        tradeId: 1,
+        entry: 105,
+        stop: 95,
+        target: 125,
+        times: {
+          h1: { decision: 3, entry: 4, exit: 8 },
+          day: { decision: 3, entry: 4, exit: 8 },
+          week: { decision: 3, entry: 4, exit: 8 },
+        },
+      },
+    ],
+    availableTimeframes: ['h1', 'day', 'week'],
+    defaultTimeframe: 'h1',
+  };
+}
+
+function caseDetail(
+  index: number,
+  model: string,
+  withTrade: boolean,
+): EpisodeReportCaseDetailView {
+  return {
+    index,
+    anchorId: `case-${index}`,
+    chartId: `trade-chart-${index}`,
+    symbol: `SYM${index}`,
+    provenanceSymbol: null,
+    provenanceLine: null,
+    questionId: `Q${index}`,
+    model,
+    modeLabel: '盲盘',
+    outcome: 'target',
+    outcomeLabel: '止盈',
+    netR: 1,
+    tone: 'positive',
+    availableTimeframes: ['h1', 'day', 'week'],
+    defaultTimeframe: 'h1',
+    planFacts: [{ label: '方向', value: '做多', tone: '' }],
+    planReasonCategoryLabel: null,
+    planReasonSummary: null,
+    planRationale: null,
+    resultFacts: [{ label: '完整交易', value: '1 笔', tone: '' }],
+    trades: withTrade
+      ? [
+          {
+            tradeId: 1,
+            direction: 'long',
+            directionLabel: '做多',
+            decisionBar: 3,
+            entryBar: 4,
+            exitBar: 8,
+            exitLabel: '止盈',
+            entryReasonCategoryLabel: null,
+            entryReasonSummary: null,
+            entryPrice: 105,
+            initialStop: 95,
+            finalStop: 95,
+            target: 125,
+            exitPrice: 125,
+            netR: 1,
+            tone: 'positive',
+          },
+        ]
+      : [],
+    actions: [],
+    process: { timingLabel: '决策 B0', hasTrace: false, events: [], checks: [] },
+  };
+}
+
+export function makeEpisodeViewData(): EpisodeReportViewData {
+  return {
+    runId: 'Episode 42',
+    generatedAt: '2026-07-21',
+    gitSha: null,
+    header: {
+      datasetChip: 'v2',
+      modelsChip: 'model-a · model-b',
+      modesChip: '盲盘',
+      costChip: '0 bps',
+      auditChip: { label: '未附加数据审计', tone: 'neutral' },
+    },
+    summarySubtitle: '2/2 完成',
+    metrics: [{ label: '胜率', value: '100%', note: '2 / 2', tone: 'positive' }],
+    configStrip: [],
+    reasonTable: { coverageLabel: '—', rows: [] },
+    modelTable: [],
+    filters: {
+      models: ['model-a', 'model-b'],
+      modes: [{ value: 'blind', label: '盲盘' }],
+      outcomes: [{ value: 'target', label: '止盈' }],
+    },
+    cases: [caseRow(0, 'model-a'), caseRow(1, 'model-b')],
+    caseDetails: [caseDetail(0, 'model-a', true), caseDetail(1, 'model-b', false)],
+    charts: [chartPayload(0), chartPayload(1)],
+    audit: { attached: false, passed: 0, total: 0, checks: [] },
+  };
+}
+
+export function makeLeaderboardViewData(): LeaderboardReportViewData {
+  return {
+    runId: 'run-1',
+    generatedAt: '2026-07-21T00:00:00.000Z',
+    title: '模型交易判断力总榜',
+    subtitle: { prefix: '2 个模型 · 1 条基线 · 10 cells', beatenLabel: '2/2' },
+    n: 3,
+    kvs: [{ label: 'DATASET', value: 'v1' }],
+    realRows: [
+      {
+        id: 'openai/gpt-5',
+        rank: 1,
+        isBaseline: false,
+        name: 'gpt-5',
+        vendor: 'openai',
+        baselineBadge: false,
+        total: '72.0',
+        delta: { tone: 'pos', text: '+5.0' },
+        judgment: { fillRatio: 0.8, text: '80.0', kind: 'j' },
+        efficiency: { fillRatio: 0.6, text: '60.0', kind: 'e' },
+        winRate: '55.0%',
+        abstainRate: '10.0%',
+        cost: '$0.0100',
+        duration: '2.0s',
+        violationRate: '0.0%',
+      },
+      {
+        id: 'anthropic/claude',
+        rank: 2,
+        isBaseline: false,
+        name: 'claude',
+        vendor: 'anthropic',
+        baselineBadge: false,
+        total: '68.0',
+        delta: { tone: 'neg', text: '−1.0' },
+        judgment: { fillRatio: 0.7, text: '70.0', kind: 'j' },
+        efficiency: { fillRatio: 0.4, text: '40.0', kind: 'e' },
+        winRate: '50.0%',
+        abstainRate: '12.0%',
+        cost: '$0.0200',
+        duration: '3.0s',
+        violationRate: '1.0%',
+      },
+    ],
+    baselineRows: [
+      {
+        id: 'baseline/buy-hold',
+        rank: null,
+        isBaseline: true,
+        name: '买入持有',
+        vendor: null,
+        baselineBadge: true,
+        total: '65.0',
+        delta: null,
+        judgment: { fillRatio: 0.65, text: '65.0', kind: 'muted' },
+        efficiency: null,
+        winRate: '—',
+        abstainRate: '—',
+        cost: '—',
+        duration: '—',
+        violationRate: '—',
+      },
+    ],
+    passLineLabel: 'BUY & HOLD 基线 · 判断分 65.0',
+    scatter: {
+      width: 460,
+      height: 320,
+      padL: 56,
+      padT: 20,
+      innerRight: 440,
+      innerBottom: 276,
+      xTicks: [{ cx: 100, label: '0' }],
+      yTicks: [{ cy: 100, label: '50' }],
+      baseline: { y: 150, label: '买入持有基线 · 65.0' },
+      dots: [
+        {
+          id: 'openai/gpt-5',
+          name: 'gpt-5',
+          cx: 200,
+          cy: 100,
+          r: 7,
+          lead: true,
+          below: false,
+          labelX: 209,
+          labelY: 92,
+          anchor: 'start',
+        },
+        {
+          id: 'anthropic/claude',
+          name: 'claude',
+          cx: 260,
+          cy: 140,
+          r: 6,
+          lead: false,
+          below: false,
+          labelX: 269,
+          labelY: 132,
+          anchor: 'start',
+        },
+      ],
+    },
+    scatterLegend: { belowLabel: '低于买入持有基线' },
+    details: {
+      'openai/gpt-5': {
+        id: 'openai/gpt-5',
+        name: 'gpt-5',
+        vendor: 'openai',
+        did: 'openai/gpt-5 · 10 cells · avg 3.0 tool-calls',
+        sections: [
+          { title: '盲盘 vs 实盘', rows: [{ label: '盲盘 判断分', value: '80.0', tone: '' }] },
+        ],
+      },
+      'anthropic/claude': {
+        id: 'anthropic/claude',
+        name: 'claude',
+        vendor: 'anthropic',
+        did: 'anthropic/claude · 10 cells · avg 2.0 tool-calls',
+        sections: [
+          { title: '盲盘 vs 实盘', rows: [{ label: '盲盘 判断分', value: '70.0', tone: '' }] },
+        ],
+      },
+      'baseline/buy-hold': {
+        id: 'baseline/buy-hold',
+        name: '买入持有',
+        vendor: 'baseline',
+        did: 'baseline/buy-hold · 10 cells · avg 0.0 tool-calls',
+        sections: [],
+      },
+    },
+    initialSelectedId: 'openai/gpt-5',
+    footer: { datasetVersion: 'v1', runId: 'run-1', generatedAt: '2026-07-21T00:00:00.000Z' },
+  };
+}
