+++ packages/bench-report-ui/test/jsdom.d.ts
@@ -1,0 +1,10 @@
+declare module 'jsdom' {
+  export interface JSDOMOptions {
+    runScripts?: 'outside-only' | 'dangerously';
+    pretendToBeVisual?: boolean;
+  }
+  export class JSDOM {
+    constructor(html?: string, options?: JSDOMOptions);
+    readonly window: Window & typeof globalThis;
+  }
+}
